+++ rspace-bench/src/main/scala/coop/rchain/rspace/bench/wide/WideBench.scala
@@ -11,7 +11,6 @@ import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.models.Par
 import coop.rchain.rholang.interpreter.accounting.Cost
 import coop.rchain.rholang.interpreter.{Interpreter, Runtime}
-import WideBench.{CoarseBenchState, FineBenchState, WideBenchState}
 import coop.rchain.shared.StoreType
 import monix.eval.Task
 import monix.execution.Scheduler
@@ -50,59 +49,27 @@ class WideBench {
   }
 }
 
-object WideBench {
+@State(Scope.Benchmark)
+class FineBenchState extends WideBenchState {
+  override lazy val runtime: Runtime = Runtime.create(dbDir, mapSize, StoreType.FineGrainedLMDB)
+}
 
-  @State(Scope.Benchmark)
-  class FineBenchState extends WideBenchState {
-    override lazy val runtime: Runtime = Runtime.create(dbDir, mapSize, StoreType.FineGrainedLMDB)
-  }
+@State(Scope.Benchmark)
+class CoarseBenchState extends WideBenchState {
+  override lazy val runtime: Runtime = Runtime.create(dbDir, mapSize)
+}
 
-  @State(Scope.Benchmark)
-  class CoarseBenchState extends WideBenchState {
-    override lazy val runtime: Runtime = Runtime.create(dbDir, mapSize)
-  }
+abstract class WideBenchState extends WideBenchBaseState {
 
-  abstract class WideBenchState {
-    val rhoSetupScriptPath: String = "/rholang/wide-setup.rho"
-    val rhoScriptSource: String    = "/rholang/wide.rho"
+  implicit def rand: Blake2b512Random = Blake2b512Random(128)
 
-    implicit val scheduler: Scheduler = Scheduler.fixedPool(name = "wide-1", poolSize = 100)
-    lazy val dbDir: Path              = Files.createTempDirectory("rchain-storage-test-")
-    val mapSize: Long                 = 1024L * 1024L * 1024L * 10L
-
-    lazy val runtime: Runtime           = Runtime.create(dbDir, mapSize)
-    implicit def rand: Blake2b512Random = Blake2b512Random(128)
-    runtime.reducer.setAvailablePhlos(Cost(Integer.MAX_VALUE)).runSyncUnsafe(1.second)
-    var setupTerm: Option[Par] = None
-    var term: Option[Par]      = None
-
-    var runTask: Task[Vector[Throwable]] = Task.now(Vector.empty)
-
-    implicit def readErrors = () => runtime.readAndClearErrorVector()
-
-    @Setup(value = Level.Iteration)
-    def doSetup(): Unit = {
-      deleteOldStorage(dbDir)
-
-      setupTerm =
-        Interpreter.buildNormalizedTerm(resourceFileReader(rhoSetupScriptPath)).runAttempt match {
-          case Right(par) => Some(par)
-          case Left(err)  => throw err
-        }
-
-      term = Interpreter.buildNormalizedTerm(resourceFileReader(rhoScriptSource)).runAttempt match {
-        case Right(par) => Some(par)
-        case Left(err)  => throw err
-      }
-
-      //make sure we always start from clean rspace
-      runtime.replaySpace.clear()
-      runtime.space.clear()
-      processErrors(Await.result(createTest(setupTerm, runtime.reducer).runAsync, Duration.Inf))
-      runTask = createTest(term, runtime.reducer)
-    }
-    @TearDown
-    def tearDown(): Unit =
-      runtime.close()
+  @Setup(value = Level.Iteration)
+  override def doSetup(): Unit = {
+    super.doSetup()
+    //make sure we always start from clean rspace
+    runtime.replaySpace.clear()
+    runtime.space.clear()
+    processErrors(Await.result(createTest(setupTerm, runtime.reducer).runAsync, Duration.Inf))
+    runTask = createTest(term, runtime.reducer)
   }
 }
