+++ rspace-bench/src/main/scala/coop/rchain/rspace/bench/wide/WideBenchBaseState.scala
@@ -1,0 +1,69 @@
+package coop.rchain.rspace.bench.wide
+
+import coop.rchain.rspace.bench._
+import coop.rchain.rholang.interpreter.ChargingReducer
+import java.io.{FileNotFoundException, InputStreamReader}
+import java.nio.file.{Files, Path}
+import java.util.concurrent.TimeUnit
+
+import coop.rchain.catscontrib.TaskContrib._
+import coop.rchain.crypto.hash.Blake2b512Random
+import coop.rchain.models.Par
+import coop.rchain.rholang.interpreter.accounting.Cost
+import coop.rchain.rholang.interpreter.{Interpreter, Runtime}
+import coop.rchain.shared.StoreType
+import monix.eval.Task
+import monix.execution.Scheduler
+import org.openjdk.jmh.annotations._
+import org.openjdk.jmh.infra.Blackhole
+
+import scala.concurrent.Await
+import scala.concurrent.duration.{Duration, _}
+
+abstract class WideBenchBaseState {
+  val rhoSetupScriptPath: String = "/rholang/wide-setup.rho"
+  val rhoScriptSource: String    = "/rholang/wide.rho"
+
+  implicit val scheduler: Scheduler = Scheduler.fixedPool(name = "wide-1", poolSize = 100)
+  lazy val dbDir: Path              = Files.createTempDirectory("rchain-wide-bench-")
+  val mapSize: Long                 = 1024L * 1024L * 1024L * 10L
+
+  lazy val runtime: Runtime = Runtime.create(dbDir, mapSize)
+  runtime.reducer.setAvailablePhlos(Cost(Integer.MAX_VALUE)).runSyncUnsafe(1.second)
+  runtime.replayReducer.setAvailablePhlos(Cost(Integer.MAX_VALUE)).runSyncUnsafe(1.second)
+
+  var setupTerm: Option[Par] = None
+  var term: Option[Par]      = None
+
+  var runTask: Task[Vector[Throwable]] = Task.now(Vector.empty)
+
+  val emptyCheckpoint = runtime.space.createCheckpoint()
+
+  implicit def readErrors = () => runtime.readAndClearErrorVector()
+
+  @Setup(value = Level.Iteration)
+  def doSetup(): Unit = {
+    deleteOldStorage(dbDir)
+
+    setupTerm =
+      Interpreter.buildNormalizedTerm(resourceFileReader(rhoSetupScriptPath)).runAttempt match {
+        case Right(par) => Some(par)
+        case Left(err)  => throw err
+      }
+
+    term = Interpreter.buildNormalizedTerm(resourceFileReader(rhoScriptSource)).runAttempt match {
+      case Right(par) => Some(par)
+      case Left(err)  => throw err
+    }
+
+    //make sure we always start from clean rspace & trie
+    runtime.replaySpace.clear()
+    runtime.replaySpace.reset(emptyCheckpoint.root)
+    runtime.space.clear()
+    runtime.space.reset(emptyCheckpoint.root)
+  }
+
+  @TearDown
+  def tearDown(): Unit =
+    runtime.close()
+}
